convert project using js to using typescript
