--- contact-management/constants.js
@@ -1,9 +1,0 @@
-const constants = {
-    VALIDATION_ERROR: 400,
-    NOT_FOUND: 404,
-    UNAUTHORIZED: 401,
-    SERVER_ERROR: 500,
-    FORBIDDEN: 403
-}
-
-export default constants
