+++ contact-management/src/constants.ts
@@ -1,0 +1,9 @@
+const constants = {
+    VALIDATION_ERROR: 400,
+    NOT_FOUND: 404,
+    UNAUTHORIZED: 401,
+    SERVER_ERROR: 500,
+    FORBIDDEN: 403
+}
+
+export default constants
