+++ contact-management/src/index.ts
@@ -1,0 +1,31 @@
+import express from 'express';
+import connectDB from './config/dbConnection.ts';
+import errorHandler from './middleware/errorsHandler.ts';
+import contactRouter from './routes/contactRoutes.ts';
+
+const app = express();
+
+const port = 5000;
+
+/* define a route BASIC
+app.get('/api/contacts', (req, res) => {
+    // res.send("Get all contacts");
+    res.status(200).json({message: 'Get all contacts'})
+})
+*/
+
+connectDB()
+
+//add json parser, to read json body from requestes 
+app.use(express.json())  //client data -x-> server ==> client data --parser--> server
+
+//Define Routes
+/*use('/url') same with requestMapping of restcontroller in springboot
+the /url will concat with require routes to become complete urls*/
+app.use('/api/contacts', contactRouter)
+
+app.use(errorHandler)
+
+app.listen(port, () => {
+    console.log(`The server is running on port ${port}`);
+})
